refactor(entity): rename entity to config
