+++ internal/cli/version.go
@@ -1,3 +1,3 @@
 package cli
 
-const version string = "v0.0.1"
+const version string = "v0.0.2"
