+++ internal/config/config.go
@@ -1,4 +1,4 @@
-package entity
+package config
 
 import (
 	"flag"
